lesson 60

+++ src/main/java/kg/attractor/movie_review_java23/dao/DirectorDao.java
@@ -15,7 +15,7 @@ public class DirectorDao {
 
     private final JdbcTemplate jdbcTemplate;
 
-    public Optional<Director> findDirectorById(int id) {
+    public Optional<Director> findDirectorById(Long id) {
         String sql = "select * from director where id = ?";
         return Optional.ofNullable(
                 DataAccessUtils.singleResult(
