+++ src/main/java/kg/attractor/movie_review_java23/service/DirectorService.java
@@ -3,5 +3,5 @@ package kg.attractor.movie_review_java23.service;
 import kg.attractor.movie_review_java23.dto.DirectorDto;
 
 public interface DirectorService {
-    DirectorDto getDirectorById(int id);
+    DirectorDto getDirectorById(Long id);
 }
